Nebraska export WIP

+++ script-engine/src/main/java/org/ihtsdo/termserver/scripting/ArchiveManager.java
@@ -32,7 +32,7 @@ import org.springframework.stereotype.Service;
 public class ArchiveManager implements RF2Constants {
 	
 	static ArchiveManager singleton;
-	
+
 	@Autowired
 	private ArchiveDataLoader archiveDataLoader;
 	
@@ -41,6 +41,7 @@ public class ArchiveManager implements RF2Constants {
 	protected TermServerScript ts;
 	protected ApplicationContext appContext;
 	public boolean allowStaleData = false;
+	public boolean loadDependencyPlusExtensionArchive = false;
 	public boolean loadEditionArchive = false;
 	public boolean populateHierarchyDepth = true;  //Term contains X needs this
 	public boolean populateReleasedFlag = false;
@@ -169,6 +170,21 @@ public class ArchiveManager implements RF2Constants {
 
 	public void loadProjectSnapshot(boolean fsnOnly) throws TermServerScriptException {
 		try {
+			if (loadDependencyPlusExtensionArchive) {
+				if (StringUtils.isEmpty(ts.getDependencyArchive())) {
+					throw new TermServerScriptException("Told to load dependency + extension but no dependency package specified");
+				} else {
+					File dependency = new File ("releases/" + ts.getDependencyArchive());
+					if (dependency.exists()) {
+						loadArchive(dependency, fsnOnly, "Snapshot", true);
+					} else {
+						throw new TermServerScriptException("Dependency Package " + dependency.getAbsolutePath() + " does not exist");
+					}
+				}
+				
+			}
+			
+			
 			//If the project specifies its a .zip file, that's another way to know we're loading an edition
 			String fileExt = ".zip";
 			if (ts.getProject().getKey().endsWith(fileExt)) {
@@ -182,8 +198,8 @@ public class ArchiveManager implements RF2Constants {
 				snapshot = new File (snapshot.getPath() + fileExt);
 			}
 			
-			//If it doesn't exist as a zip file locally either, we can try downloading it from S3
 			if (!snapshot.exists()) {
+				//If it doesn't exist as a zip file locally either, we can try downloading it from S3
 				try {
 					String cwd = new File("").getAbsolutePath();
 					TermServerScript.info(snapshot + " not found locally in " + cwd + ", attempting to downlod from S3.");
@@ -260,17 +276,22 @@ public class ArchiveManager implements RF2Constants {
 							Boolean isReleased = loadEditionArchive ? true : null;
 							loadArchive(snapshot, fsnOnly, "Snapshot", isReleased);
 						} catch (Exception e) {
-							TermServerScript.error ("Non-viable snapshot encountered (Exception: " + e.getMessage()  +").  Deleting " + snapshot + "...", e);
-							try {
-								if (snapshot.isFile()) {
-									snapshot.delete();
-								} else if (snapshot.isDirectory()) {
-									FileUtils.deleteDirectory(snapshot);
-								} else {
-									throw new TermServerScriptException (snapshot + " is neither file nor directory.");
+							TermServerScript.error ("Non-viable snapshot encountered (Exception: " + e.getMessage()  +").", e);
+							if (!snapshot.getName().startsWith("releases/")) {
+								TermServerScript.info ("Deleting " + snapshot);
+								try {
+									if (snapshot.isFile()) {
+										snapshot.delete();
+									} else if (snapshot.isDirectory()) {
+										FileUtils.deleteDirectory(snapshot);
+									} else {
+										throw new TermServerScriptException (snapshot + " is neither file nor directory.");
+									}
+								} catch (Exception e2) {
+									TermServerScript.warn("Failed to delete snapshot " + snapshot + " due to " + e2);
 								}
-							} catch (Exception e2) {
-								TermServerScript.warn("Failed to delete snapshot " + snapshot + " due to " + e2);
+							} else {
+								TermServerScript.info ("Not " + snapshot + " as it's a release.");
 							}
 							releasedFlagPopulated = false;
 							throw new TermServerScriptException("Non-viable snapshot detected",e);
